Modify DELAY_SAMPLE and SAMPLE_CYCLE_8
Add instructions to DELAY_SAMPLE macro. The macro now uses the
SAMPLING_CONFIG_CYCLE_BTWN_SAMPLE, subtracts 7 and stores the result
in R0. Further it gets in a loop to cause a delay of R0 (value in R0 )
cycles.

The subtraction of 7 cycles is because there are separate delays in
the SAMPLE_CYCLE itself. Thus it turns out that the minimum delay
between consecutive samples can be 7 and thus least value of
SAMPLING_CONFIG_CYCLE_BTWN_SAMPLE can be 7. The value in the
SAMPLING_CONFIG_CYCLE_BTWN_SAMPLE should be exact equal to the delay
required between 2 consecutive sample, ignoring all delays caused by
Clock generation or anything. That will be managed in the code itself.

SAMPLE_CYCLE_8 macro was also changed. It is still incomplete and was
changed to avoid compilation errors due to change in DELAY_SAMPLE
macro.