ADD BLINK and add details to comment
Add a BLINK macro for debugging process while developing the code. It
uses the does nothing more than toggling the CLK_PIN.
Also added more details to the comments that describe the DELAY_SAMPLE
macro.The Value given in the SAMPLING_CONFIG_CYCLE_BTWN_SAMPLE should
always be odd to work correctly, or else it will get stuck in an
infinite loop.