Change config data storage
Earlier SAMPLING_CONFIG was being used to srote data. Now
SAMPLING_CONFIG_0 and SAMPLING_CONFIG_1 registers are being used to
store config data. This gives us more space to store data, if needed
in future. SAMPLING_CONFIG_START is the firts register and the
SAMPLING_CONFIG_LENGTH is the length of the config data in bytes.