Change configuration of INTC and R31 write instruction
The pru0 code that is used for testing and debugging purpose, was
made to generate Interrupt INT_P0_to_P1, but the INTC was not
configured and the value that was being written to R31 was wrong.
As a result, the interrupt was not being generated.

This commit corrects those mistakes and also increases the gap
between consecutive interrupts generation. This has been tested and
is working.