Add CHECK_INT_LOOP and rewrite MANAGE_INTERRUPT macro
The CHECK_INT_LOOP macro keeps polling for the occurrence of
INT_P0_to_P1 interrupt and once the interrupt status is positive,
program control moves to the MANAGE_INTERRUPT macro.

The MANAGE_INTERRUPT as of now, is incomplete. It just clears the
status of INT_P0_to_P1 interrupt and then calls the BLINK macro.

There may be significant changes in the CHECK_INT_LOOP macro in the
future, as this approach of detecting macros consume 3 cycles.