Modify resource table file for pru0
Pru0 source code was using resource_table_pru1.h, which was actually
meant for pru1 source code. This was still working as the two
resource files were identical and essentially empty. This has been
corrected so as to not cause conflicts in future.

Further, resource_table_pru0.h included pru_defs.h, which is not a
part of this firmware. This was not producing error as the
resource_table_pru0.h was itself not being used. The include statement
for pru_defs.h was therefore removed.