Add interrupt form PRU1 to PRU0
After sampling process, PRU1 will have the sampled data. PRU1 will transfer
this data to one of the Banks in scratch pad and then generate the system
interrupt INT_P1_to_P0 to interrupt PRU0 and make it read data and proceed
further.

In this commit, the interrupt INT_P1_to_P0 was added. Now on loading the
firmware to the PRUs, PRU1 writes sampling_config data to the shared RAM
and waits for a message from linux userspace that can be send using
/dev/rpmsg-pru30 char device file. On receiving the message, PRU1 interrupts
PRU0. PRU0 further sampling config data from shared RAM and then executes
BLINK. After executing BLINK, it generates an interrupt back to PRU1. As soon
as PRU1 gets this interrupt, it sends an "INTERRUPT" message to the linux
usersapce that can be seen by reading the /dev/rpmsg-pru30 file.

In total, the changes made in this commit check the overall interrupt and
communication settings.

Changes made -
common_pru_defs.h : Added constant that contains the value that needs to be
written to R31 in order to generate INT_P1_to_P0

main_pru0.c : Added code that will be polling the status of INT_P0_to_P1
interrupt and if detected, sends message to linux userspace.

main_pru1_def.asm : Add macro TRANSFER_AND_TELL. The macro is incomplete
and as of now, generates INT_P1_to_P0 interrupt.