Add PRU1 to PRU0 data related constants
The PRU1 sends sampled data to PRU0. The constants added in this
commit are related to this data.
DATA_SIZE is the size of the data transfered by PRU1 for each
interrupt to PRU0.
DATA_START_REGISTER and DATA_START_REGISTER_NUMBER are related to the
first register, from where the sampled data starts getting saved
in PRU0