Modify main_pru0.c to read from the scratch pad
The pru0 source code file was modified to make pru0 intake the 44
bytes of data from the 3 scratch pad banks. On every interrupt
generated by PRU1, PRU0 intakes data from one of the banks.
It starts from bank0 on first interrupt and then advancing to the
next bank for next interrupt until bank2, from where it gets back to
bank0.
This data is saved in a structure in the PRU0.