Update 'Register usage' in the main_pru1.c file
The register usage section that describes and declares usage of
register in PRU1 was updated.
Changes were made to reserve a space of 88 bytes to save input sample
data. This was because the maximum data that can be read by PRU0
from scratch pad bank at a time is 44 bytes. So the idea is to send
88 bytes of data from PRU1's side in a single step and code PRU0 to
read this data in 2 steps of 44 bytes each.