Add support to set sampling configuration from userspace
Till now the configuration settings for the sampling process were
being set by PRU0. PRU0 did this by  writing predefined configuration
data into the shared memory and then generating interrupt to PRU1.

This commit adds support to set these configuration settings from the
userspace. To set the settings. Kernel needs to send 2 separate
messages containing 32 bit configuration data in each. The data in
the first message will be used as SAMPLING_CONFIG_0 and the
data in the second message will be used as SAMPLING_CONFIG_1.
SAMPLING_CONFIG_0 and SAMPLING_CONFIG_1 are defined in PRU1's source
code. The PRU also echoes back the data it receives.

This can only be done once, when the PRUs are rebooted. Once PRU1
starts sampling, it will continue sampling with same configuration
data, until the PRU is not rebooted.

Internally, PRU0 just takes the data sent to it by the kernel, writes
it into appropriate position in the shared memory and then generates
the interrupt.