Modify SAMPLE_CYCLE_8 to switch to new configuration faster.
In the previous versions, the changing of configuration data was
supported but PRU1 adopting the new configuration data was very slow.
This was because it has to wait for a sampling_cycle of 44bytes to
complete.

This commit removes this limitation and makes changes to make this
transition to new sampling configuration data faster. This was done
by using CHECK_INT macro after each sample is taken in SAMPLE_CYCLE_8
macro.

Changes:

SAMPLE_CYCLE_8 macro : call to CHECK_INT macro was included after
every sample is taken by TAKE_SAMPLE_8 macro, but the last sample.
CHECK_INT was not added after the last sample to maintain
a consistent delay between all the samples.