Add comments to main_pru0.c
main_pru0.c is the source file for PRU0's code. Until this commit,
very little attention was being paid to the readability of PRU0's
code.

This commit adds comments to main_pru0.c file, explaining various
tasks and functions that are being executed by the PRU.

NO C INSTRUCTION WAS CHANGED/MODIFIED IN THIS COMMIT. THE ONLY CHANGE
THIS COMMIT ADDS TO main_pru0.c IS THE ADDITION OF COMMENTS.