Reformat main_pru0.c code.
The source code for main_pru0.c was a bit messed up, making it
difficult to read it.

This commit reformats the code, so as to look better and easy to read
and understand.

There were some minor modifications in some of the instructions, like
combining the declaration and initialization steps into one single
step.