Start developing driver for BeagleScope
By this time, the firmware is in a stage that it can take the input
data and send it to the kernel using rpmsg. Working on driver can now
be started.

The idea is to first make one single driver, straight from iio
subsystem to the rpmsg and remoteproc drivers to handle the pru.
Therefore, as development will go further, this driver will first
increase in size, and then will be  broken down to smaller specific
drivers. And finally this driver will disappear leaving behind other
drivers depending upon the role they perform.