Add 'install' target to Makefile
To help in the development process, and to make development process
faster, add an 'install' target to Makefile.

The install target builds the driver and copies into the
kernel/drivers/testing/ directory and executes 'depmod -a' command.
This is necessary to use the 'modprobe' command which would manage
driver dependencies itself.