Add the missing devm_iio_device_free() functions
The driver in its probe function allocates space for the iio device
data using devm_iio_device_alloc() function.

The device was missing the function that deallocates the allocated
space, when any error would occur or when the device is removes.
This is done by the using devm_iio_device_free().