Add fifo and callback method to get data from pru
This commit adds kfifo to the beaglescope_state structure and a
callback method - beaglescope_driver_cb, to get data from the pru.

The kernel fifo will provide the buffer to save data into. Data can
then be read by other methods from this fifo.

The callback method is invoked by the virtio_rpmsg_bus whenever there
is some data to be read sent by the pru. The method stores the data,
that comes from the PRUs, into the kfifo buffer.