Add support to prevent writing to kfifo when it is full
The data_fifo used to buffer data from PRUs is limited in size. There
need to be a way to prevent writing to the buffer when it is full.

This commit adds support to prevent the writing any data further to
the kfifo when it is full. This is done by -

The fifo is divided into 'MAX_BLOCKS_IN_FIFO' blocks, each of
'FIFO_BLOCK_SIZE' bytes.
The 'data_idx' member of the beaglescope_state structure saves the
current index of the fifo block and checks its value before writing to
the fifo. If the data_idx is at its last index ie
data_idx == MAX_BLOCKS_IN_FIFO, no data is further written to the fifo