Rename object rpmsg_dev to rpdev in probe and remove functions
The object of the rpmsg_channle structure that was being used in
the probe and remove function of the driver was 'rpmsg_dev'.
The structure is renamed to 'rpdev' to maintain consistency.

'rpdev' is what mostly used as the name to the object of rpmsg_channel
structure and thus makes it easy to understand the code.