Add test.patch to patch the code for testing purpose
When the beaglescope_driver.c is patched using test.patch, it is
modified for testing purposes.

This modifies the driver code, to send an initial message to the pru
in driver's probe function. It further adds a dev_err message to the
callback function to view the view the arrival of the new message from
the pru side, in the dmesg.