Remove DELAY_IMMEDIATE_2n macro and add DELAY_2 macro
DELAY_IMMEDIATE_2n macro was used to cause a delay of 'cycles_2n'
cycles, where 'cycles_2n' is a multiple of 2. The macro can be used to
cause any delay greater than 2, and multiple of 2. The macro uses 4
instructions to achieve this. Further since its being used multiple
times in the SAMPLE_CYCLE_8 macro, it results in a bit lengthy code.

Since the DELAY_IMMEDIATE_2n was being used to cause a delay of only
2 cycles, this commit removes DELAY_IMMEDIATE_2n macro and adds
DELAY_2 macro in place of it. The DELAY_2 macro causes a delay for
only for 2 cycles, without those extra instructions.

DELAY_2 macro uses just 2 NOP instructions to cause 2 cycle delay.
This somewhat decreases the length of the code, as its replaces the
DELAY_IMMEDIATE_2n macro in the SAMPLE_CYCLE_8, where the former was
being used 44 times. This will help in further development.