Add 'THE_DELAY' macro
'THE_DELAY' macro is yet another delay macro, that can cause a
configurable delay, with a special point that the least delay it can
be used for is 1 cycle. It can also be used for higher, odd numbered
delays.

The macro is named 'THE_DELAY' as it might replace all other
configurable delay macros in the code