Modify Register usage to increase the size of configurable data
The 'Register Usage' section in the main_pru1.c file defines the use
of all the PRU1 registers for different purposes. The BYTE_N registers
are used to temporarily store sampling data. The SAMPLING_CONFIG_N
registers are used to store the 'configuration data that is used for
sampling process.

This commit modifies this section to use R29 register as a
SAMPLING_CONFIG_N register removing it from the BYTE_N registers. Thus
the size of the configuration data also increases from  8 bytes to
3 * 4 (12) bytes