Add CYCLE_BEFORE_SAMPLE and CYCLE_AFTER_SAMPLE constants
In earlier versions, the configuration data is divided into 2 parts.
The CYCLE_BTWN_SAMPLE and MISC_CONFIG_DATA part. The first one
containing the number of delay cycles between each sample, and the
second one containing miscellaneous configuration data like the
SAMPLING_START_BIT or the SAMPLING_WIDTH.

This commit add 2 more constants CYCLE_BEFORE_SAMPLE and
CYCLE_AFTER_SAMPLE that use the SAMPLING_CONFIG_1 the lower word and
the higher word of the register respectively.

The CYCLE_BEFORE_SAMPLE value is the number of cycles to be delayed
after the CLK_PIN is high, but before the sample is take.
The CYCLE_AFTER_SAMPLE value is the number of cycles to be delayed
after the sample is taken, but before the CLK_PIN is back to low.

This commit is basically the first step to solve the #1 issue.