Modify SAMPLE_CYCLE_8 to use .loop,.break and .endloop directives
SAMPLE_CYCLE_8 macro is used to take a complete cycle of 44 samples
and save them in the register files. Call to thus macro in the 'main'
is followed by the call to TRANSFER_AND_TELL macro and CHECK_INT/JMP
to transfer data to PRU0 or to restart the loop.

This commit modifies the structure of the SAMPLE_CYCLE_8 and 'main'
using the .loop, .break and .endloop assembler directives to use the
code in loop. Thus now, calling SAMPLE_CYCLE_8 macro three times in
'main', each time for a different scratch pad bank is not required.