Change value of MISC_CONFIG_DATA
After adding support for extra 32 bits of configuration data, the
data in the SAMPLING_CONFIG_1 was reserved for
CYCLE_BEFORE/AFTER_SAMPLE, which was earlier being used for
MISC_CONFIG_DATA.

As a result of this change, MISC_CONFIG_DATA was to be defined as
SAMPLING_CONFIG_2, but was mistakenly left unchanged. This commit
corrects this mistake.