Add support for RAW_READ and BLOCK_READ read modes
In the earlier versions of code, the PRUs could only sample data
continuously and then send the sampled data in blocks of 44bytes

This commit adds support for different reading modes. The two
supported modes are READ_RAW and BLOCK_READ.

In READ_RAW mode, the PRUs sample a single data byte of data and send
it to the userspace.

In BLOCK_READ mode, the PRUs sample continuously and send the data to
the kernel is blocks of 44bytes of data.

The READ_MODE is configured by the 1st byte of the 3rd configuration
message that is sent to the PRUs in order to configure them