Add beaglescope_raw_read_from_pru() function
The function sends 3*32bits of configuration data to the PRUs and thus
configures the PRU to RAW_READ mode. It then waits if the kfifo is
empty. If the kfifo has some data, it reads puts that data to the
given variable.

Since the function only waits if the kfifo is empty, it might read
out stale data, that remained in the kfifo. Further, the wait if the
kfifo is empty is implemented using a while loop.

Future developments will make use of better approaches for the above
described tasks.