IMP : Add new line chars at the end of debug messages
Man ! I need to get into habit of adding 'new-line' characters at the
end of the debug message. Really !! This carelessness and mistake can
be disastrous.

As this single mistake can lead to heating up of the BBB board, which
further can explode. This explosion would produce enough heat to cause
a nuclear fission.
If it happens, according the Einstein's relativity, using equation
E=mc^2, the energy then produced will be 3*10^17 Joules. Which,
if the dark energy exists, can result in opening up of a wormhole, or
collapsing of the space-time, and resulting into
space-time-singularity.
Singularity can only be around an object with high gravitational pull
, such as a blackhole. Thus this can also produce a black hole, that
may eat up the whole universe.

But wait, all this can be prevented, if I develop the habit of putting
a 'new-line' character at the end of every kernel log message soon.