Modify SAMPLE_CYCLE_8 macro and 'Register usage' section
The SAMPLE_CYCLE_8 macro in main_pru1_def.asm file is used to take
data input and transfer it to the scratch pad banks. Until this commit
the macro used repetitive use of sample taking steps. This made the
code very big and addition of any extra code was resulting in
compilation errors as the resultant code was too big for the IRAM in
the PRUs.

This commit modifies the working of SAMPLE_CYCLE_8 so as to use MVI
instruction in a loop rather than using repetitive sample taking
steps, to take input data and send it transfer it to the scratch pad
banks.

The SAMPLE_CYCLE_8, as of now in not fully optimized in terms of delay
between consecutive samples and can only transfer data to one of the
scratch pad banks.

The 'Register usage' section was updated to reserve registers to be
used by the modified SAMPLE_CYCLE_8 macro.