Replace use of DELAY_SAMPLE macro by THE_DELAY macro
The DELAY_SAMPLE macro was used to produce a delay between consecutive
samples. The input give to this macro was a register containing the
exact number of cycles that was required to be delayed between
consecutive sample. By 'exact' it is meant that the number of cycles
without subtracting any other delay caused by the clock toggling
instructions or any other delays.

THE_DELAY macro is different and more useful. The input that this
macro takes is a register that contains the raw number of delays
and the actual delay between consecutive samples needs to be
calculated adding delays by other instructions in the sample taking
step, inside the SAMPLE_CYCLE_8 macro.
Other useful feature of THE_DELAY macro is that its least count is 1
cycle.
Also note that THE_DELAY macro can only accept an odd number of delay
cycles.

As of now, the least latency between two consecutive sample that this
code can produce is of 6 cycles, that is, 30 nano seconds.