Add support for configurable clock duty cycle.
In the previous version, the sampling was done by using the
TAKE_SAMPLE_8 macro. The TAKE_SAMPLE_8 macro performed following
steps :

- Toggle clock pin and thus make it HIGH
- Take the input from GPIs and stores it in one of the register
- Toggle clock pin and thus changes its state to LOW

Thus there was no way to make the clock duty cycle configurable.

This commit adds support for configurable duty cycle. Thus this commit
finally closes #1.

The sampling procedure, that is followed inside the SAMPLE_CYCLE_8
macro is as :

- Toggle clock and thus make it HIGH
- Wait for 'CYCLE_BEFORE_SAMPLE' number of cycles
- Take the input sample and store it in one of the registers
- Wait for 'CYCLE_AFTER_SAMPLE' number of cycles
- Toggle clock and thus make it LOW
- Waits for 'CYCLE_BTWN_SAMPLE' number of cycles
- Checks if there is any interrupt using CHECK_INT macro or is used for
  some house keeping instruction
- Instruction to transfer data, or for the counter, or to reset values
- 2nd Instruction to transfer data or any such task

The value of the CYCLE_BTWN_SAMPLE, CYCLE_BEFORE_SAMPLE and
CYCLE_AFTER_SAMPLE comes from the configuration data.

The above steps are performed in loop to allow continuous sampling