Add support to use all the 3 scratch pad banks
After the modification in the SAMPLE_CYCLE_8 macro to use MVI
instruction to take data sample, the support to use all the 3 scratch
pad banks was removed. This was to make things simple as more
work was needed to be done.

This commit adds support to use all the 3 scratch pad banks in a loop
starting from SP_BANK_0 to SP_BANK_2.

This is basically done by using the previous code that supported only
one bank, and copy-pasting it to support all the tree banks.

SAMPLE_CYCLE_8 macro is the only macro that was modified for this
purpose.