Change configuration data from 3*32 bit message to 1*96 bit message
Until this commit, the configuration data to the PRUs was required to
be send as 3 different messages, each 32 bit, that is, 4 byte long.
Thus making a total of 12 bytes of configuration data.

This commit modifies main_pru0.c to make it accept configuration data
in one single 12 byte long message.

The configuration data is saved in a 12 byte long buffer as :
[ CYCLES_BTWN_SAMPLE CYCLE_BEFORE_SAMPLE CYCLE_AFTER_SAMPLE MISC_CONFIG_DATA]

where
- CYCLES_BTWN_SAMPLE is 4 byte long
- CYCLE_BEFORE_SAMPLE is 2 byte long
- CYCLE_AFTER_SAMPLE is 2 byte long
and
- MISC_CONFIG_DATA is 4 byte long