Add support to start and stop sampling a single byte message
This commit adds support to start and stop the sampling process by the
PRU1 by sending a single byte message from the kernel space. The
older configuration data remains preserved along this switching.

Sending a 12 byte long message to just start and stop the PRUs is
not efficient. Therefore this switch was added to the PRUs.