Change RPMsg channel name created by the PRU firmware
The PRUs create an RPMsg channel to transfer data to the ARM. The name
of this RPMsg channel decides the driver that will be probed for the
PRUs.

As now the work to couple PRU firmware and the driver is to be done,
the channel name of the firmware is changed.