Add log_debug macro to beaglescope_driver
I had some hard time getting the pr_debug and dev_debug methods print
debug message to the kernel log. That is probably because of their
internal functioning and that they have some default level fixed.

On the other side, the format to use printk() is a bit useless for
small messages.

This commit adds a very simple macro called log_debug to print debug
message to the kernel logs easily using printk().