Modify beaglescope_raw_read_from_pru() to send only one config message
In the earlier versions of the source code, the PRU required 3
separate messages from the kernel to set its sampling configuration.

The latest modifications in the PRU firmware now take only one message
from the kernel and uses this message as its config info. Though this
message still comprises of 3 different messages in form of 3 variables
in a u32 array. But the data is send as a single message.

This commit modifies the beaglecope_raw_read_from_pru() method to
send only one message to the PRUs in order to configure them to
READ_RAW mode.