Remove kfifo use
For now, there is no buffer required, as the primary goal is to
implement read_raw method for the driver. Sine read_raw method needs
only one data value, no buffer is required.
Even if a buffer is required in future, IIO ring buffer will probably
be used.

This commit removes the use and initialization of the kfifo that was
being used as a buffer for the data.