Add beaglescope_read_raw() associated to read_raw field in iio_info struct
The IIO subsystem creates various sysfs entries to expose different
functionalities of the device. The sysfs entries are created on the
channel specifications of the device, which are specified inside the
iio_chan_spec structure. One of such sysfs is in_voltage_raw.

This entry is used to acquire one datum from the device that measures
voltage. The sysfs entry is bound to a function inside the driver
pointed by read_raw field of the iio_info structure. For this driver
it is the beaglescope_read_raw().

The beaglescope_read_raw() is uses beaglescope_read_raw_from_pru() to
acquire one sample data from the PRUs.