Add beaglescope_block_read_from_pru() in the beaglescope_driver.c
The function sends configuration data to the PRU over rpmsg_bus, so as
to configure the PRUs to BLOCK_READ mode and start sampling. This is
done be sending an array of 3*32bits data in a single shot, using
rpmsg_send() just once. Thus the message to pru is of 96bits.

For now, the function just configures the PRU with random data, but in
BLOCK_READ  mode. Further developments will add support of sending
configuration data according to the beaglescope_state and selected
sampling frequency.

BLOCK_READ and RAW_READ are the two read modes added to the firmware
https://github.com/ZeekHuge/BeagleScope/commit/bbd22b00ad946f8f9a306a8a9e0a582bddd63b0b