Add beaglescope_stop_sampling_pru to stop sampling process immediately
Call to this function stops the sampling process done by the PRU
immediately. This is done by sending a single byte - 0x00 to the PRUs.
The sampling process can then be started back with the same
configuration by sending single byte - 0x01, but this support has not
yet been added to the driver.

This function basically deploys features added to the firmware in this
commit :
https://github.com/ZeekHuge/BeagleScope/commit/ad6f6c39d8976eb310dad644ba0452de7f9823c1