Add INDIO_DIRECT_MODE and INDIO_BUFFER_SOFTWARE as iio device modes.
Support to sample a single instantaneous value for the read raw mode
was already done in earlier commits. Work now is being done to add
buffered capture.

The commit adds INDIO_DIRECT_MODE and INDIO_BUFFER_SOFTWARE as the
read modes for the iio device. This is important as this is what tells
the IIO subsystem about the features that are supported by the device.

INDIO_DIRECT_MODE basically tells that the device supports
instantaneous read of a single data value.

INDIO_BUFFER_SOFTWARE tells the IIO subsystem that the device driver
supports storing data to a kernel (software) buffer.