Modify the beaglescope_driver_cb() to support READ_RAW and BLOCK_READ reading modes
The beaglescope_driver_cb() is the callback method that get called
whenever there is message for the rpmsg_channel associated to this
driver.
This callback is responsible to then dispatch the message to the write
place and use it correctly.

This commit modifies the callback so as to support BLOCK_READ and
RAW_READ modes. The two modes are different as the BLOCK_READ needs
iio_buffer while RAW_READ just need a single variable to cache the
instantaneous values.

This commit is a step forward in adding buffer capture support.