Modify beaglescope_buffer_postenable and predisable to add buffer capture
The commit finally adds buffer capture mode by connecting the parts
that were developed in previous commits.

This is basically done by calling beaglescope_block_read_from_pru()
and beaglescope_stop_sampling_pru() from buffer_posteneable and
predisable functions.

___

The flow of control for this driver is as :

Whenever the buffers are enabled by the user, the IIO subsystem
invokes the buffer_postenable function just after that.

The postenable function then invokes beaglescope_block_read_from_pru()

block_read function then configures the PRU to block reading mode,
with other necessary config data. For now, this config data is just
predefined. Further development will add support to use config data
according to the desired result.

After the PRUs are configured to BLOCK_READ mode and the sampling
starts, the PRUs send data to the driver in blocks of 44bytes. The
beaglescope_driver_cb callback is invoked each time there is message
from the PRUs. This method then uses this data according to the
read_mode.
If the read_mode == RAW_READ, the data is a single sample and is
cached.
If the read_mode != RAW_READ, it can only be BLOCK_READ, in which case
the data is 44bytes block and  is pushed to the IIO buffers, in a loop,
8 bytes each time.
___