Remove unused variables out of beaglescope_state structure
The beaglescope_state structure it to store the current state of the
beaglescope device. This structure has all the data about the current
instance of the device that can be used by different functions.

This commit removes:
1. data_kfifo
2. data_idx
and
3. data_length[MAX_BLOCKS_IN_FIFO]
these variables/structures were being used in the older versions of
the driver. Since the driver now uses iio_buffer, these kfifo related
variable are not required