Add pru_config and sampling_frequency to the beaglescope_state struct
beaglescope_state structure is to keep save data about the current
instance of the device.

This commit adds:
1. pru_config
and
2. sampling_frequency
to the beaglescope_state structure.
The driver as of now does not support to set the PRUs at variable
frequency based on the input from the user. This is the first step in
adding that support. Ultimately, the pru_config will hold the
configuration data to configure the PRUs based on the
sampling_frequency and read mode. Thus, the pru_config is actually a
3*32bit long buffer.
After some more development, sampling_frequency will save the value of
sampling frequency, PRU is currently set to. The value will actually
be based on the input from the user.