Add set_beaglescope_sampling_frequency() to the driver
The PRUs need to have some sampling frequency to actually do a
buffered capture. In the earlier versions of code, having support for
buffered capture, the PRU was actually being configured to a
predefined sampling frequency. This commit is a step forward in adding
variable sampling frequency support.

The PRU does not understands frequencies, it understands the number of
cycles it has to delay. Each cycle is of 5 nano seconds. And hence the
time period of the sampling will be an Arithmetic progression with 5
as common-difference. This makes the attainable frequencies and
Harmonic progression, and harmonic progressions have no pattern.

The set_beaglescope_sampling_frequency() takes the sampling frequency,
does the necessary math and get the number of PRU cycles to delay. It
then fills in the pru_config buffer of the current beaglescope_state
structure. Note that only the first 2*32bits of the buffer are
used and configured by this function as only this part of the config
buffer is related to the number of delays.
Since, not all frequencies are attainable, the function gets out the
closest possible value, greater than or equal to the give frequency.
The function then saves this value as the sampling_frequency in the
current beaglescope_state structure.

The function is not used yet, it is just defined for now.