Add set/get_beaglescope_read_mode()
In the earlier versions of the code, the read mode of the current
beaglescope state was being monitored by a variable from the
beaglescope_state structure.
To make code more efficient and so as to be able to get the
appropriate pru-configuration data, this commit adds 2 more functions
to the drivers:
1. set_beaglescope_read_mode()
	This function basically just set or clears the bit associated
	to the read mode, in the pru_config buffer. It also sets the
	enable bit in the pru_config buffer.
2. get_beaglescope_read_mode()
	This function reads back the bit associated to the read mode,
	inside the pru_config buffer, and returns its value.

Thus these two set of functions provide a way to set and get the
read mode of the of the current beaglescope state, while configuring
the pru_config buffer on the go.

The driver are not used for now, they have only been defined in the
driver.