Add sampling_frequency read and write support
This commit adds support to read an write sampling frequency for
buffered capture, from the userspace.
The modifications added, use this given sampling frequency from and
generate the pru_configuration data and saves it in the pru_config
buffer.

In order to get the sampling_frequency from from the user, this commit
also adds write_raw function. But it is only used for
sampling_frequency and nothing else.

As of now, this configuration data is not being used. This is because
the driver yet do not uses 'beaglescope_read_from_pru()' to read.
Further development will use the other defined-but-unused methods too.