Changed the sampled_data buffer from structure to array
Data, after being sampled by PRU1 is transfered to one of the scratch
pad banks and then PRU0 is interrupted. PRU0 then transfer this data
from the bank to a buffer in its own memory region.

This buffer, in the earlier versions was in the forms of a structure
with a uint8_t-type array, of 'DATA_SIZE' length, as the only element.

This commit removes that structure and uses only the uint8-type array,
of length 'DATA_SIZE' as the buffer. This is to simplify the code.