Start using the pru_rpmsg_send_large_buffer() in the code
Previous commits added definition of the
pru_rpmsg_send_large_buffer() to the source code, but did not actually
use it in the code.

This commit uses this function by replacing call to pru_rpmsd_send()
with the call to pru_rpmsg_send_large_buffer() in the BLOCK_READ
read_mode.