Modify log_debug() macro to provide better debugging aid
In this commit, the log_debug macro, that was using printk() to print
temporary debug messages was modified to better serve the purpose.

Earlier it was taking in a string as the message and using it as the
printk-message. This commit modifies it and make it print the module
name and the function from which it is called.