Add parallel_interface.c parallel_interface.h and pi-bus.c to the source.
Since I am learning along, I have been doing a lot of experimenting and testing
on other wip branches in order to begin the development of the
parallel interface APIs.
This commit adds following source files:

1. pi-bus.c : The platform driver to provide a way to the core driver to talk
to the bus.

2. parallel_interface.c : The core bus driver, that will allow device driver
to talk to the parallel bus with a number of APIs.

3. parallel_interface.h : The header file that the core bus driver, device
driver and platform driver needs to include.

As of now, these source files contain some basic code, just enough to make them
modprobe-able and rmmod-able.