Add rpmsg-client support to the pi-bus driver.
The pi-bus driver is the platform specific driver. The platform - beaglebone
black is what this pi-bus driver is specifically for and is meant to use PRU as
a parallel bus.

To communicate to the PRUs from the Linux side, one need to create an rpmsg
channel that would further communicate to the PRUs.

This driver therefore needs to add support for the rpmsg-client driver. Further
development would use this driver to communicate data to the PRU firmware.