Add support in the pi-bus driver to serve as a platform-driver.
The pi-bus driver is the platform specific driver that would be different for
different platforms. This commit adds support to make the pi-bus driver serve
as a platform-driver.

The driver now registers itself as a platform driver with an of-device-id and
probe and remove function to be invoked when a matching platform device is
found.

The platform registration process is done inside the rpmsg-probe method. This
means that the driver can only serve as the platform driver when the PRU is
booted with correct firmware.