Add code in the parallel_interface driver to register itself as a bus driver
The parallel_interface driver is the core bus driver. As a result, it needs to
register itself to the kernel as a bus driver.

This commit adds support in the parallel_interface driver to register itself
as a bus driver by using an object of type 'struct bus_type'.

After registration, the driver will be ready to add support to register devices
and device-drivers. Further, once the registration is done, associated sysfs
entry will also be created.

Associated sysfs entry :

root@beaglebone:~# tree /sys/bus/parallel_interface/
/sys/bus/parallel_interface/
|-- devices
|-- drivers
|-- drivers_autoprobe
|-- drivers_probe
`-- uevent