Add pi_bus_host structure in parallel_interface.h header file
This commit adds pi_bus_host structure definition to be used by the drivers
that include the parallel_interface.h header file. The pi_bus_host structure
represents the host device that is present on the system, this driver is being
used on. All devices that are connected to this bus will actually be child
devices to the pi_bus_host.

Further developments will add APIs to register bus-host device.