Add API support to register and unregister pi-bus host device
This commits adds supports and declares::
1. pi_core_register_host()
and
2. pi_core_unregister_host()
APIs to register and unregister a host device.

The platform-specific driver looks for the platform bus host device and uses
the pi_core_register_host() API to register the device with the pi bus core
driver. This host device is the principal device on the bus and all other
devices connected on the bus will be child to this device.

A successful call to device would result in a host-device named 'pi-0' on the
pi bus.

NOTE: The API just supports one host device Calling the register API more than
once would result in unexpected results.